Migrate pure-logic tests from androidTest to JVM unit tests
