+++ ui-lib/src/test/java/co/electriccoin/zcash/ui/screen/send/ext/ZecSendExtTest.kt
@@ -1,7 +1,6 @@
 package co.electriccoin.zcash.ui.screen.send.ext
 
 import androidx.compose.runtime.saveable.SaverScope
-import androidx.test.filters.SmallTest
 import cash.z.ecc.android.sdk.model.ZecSend
 import cash.z.ecc.sdk.fixture.ZecSendFixture
 import kotlinx.coroutines.test.runTest
@@ -10,7 +9,6 @@ import kotlin.test.assertEquals
 
 class ZecSendExtTest {
     @Test
-    @SmallTest
     fun round_trip() =
         runTest {
             val original = ZecSendFixture.new()
@@ -27,7 +25,6 @@ class ZecSendExtTest {
         }
 
     @Test
-    @SmallTest
     fun restore_empty() {
         val restored = ZecSend.Saver.restore(emptyList<Any?>())
         assertEquals(null, restored)
